refactor(monorepo): extract app/shared-ui workspace [A3] (#20)
A3 of the mobile-platform plan (ADR 0015). Hoists the 16 shadcn-derived
UI primitives out of app/web/ into a new app/shared-ui/ workspace package
so the upcoming app/mobile/ entry can mount the same component library
without duplication.

This is the "Option 1 / minimum scope" framing discussed earlier:
shared-ui starts with primitives only. Business components
(sessions/, integrations/, backup/, etc.) stay in app/web/ for now;
they migrate case-by-case once mobile actually needs them, following
YAGNI.

## What moved (16 primitives, all `git mv` to preserve history)

app/shared-ui/src/primitives/:
  badge, button, code, command, dialog, dropdown-menu, input, label,
  scroll-area, select, separator, sonner, switch, tabs, textarea,
  tooltip.

app/web/src/components/ui/ removed (empty after move).

## How web finds them — alias `@/components/ui/*`

- `app/web/tsconfig.app.json`:
  - new path mapping `"@/components/ui/*": ["../shared-ui/src/primitives/*"]`
    placed before the generic `@/*` so it wins
  - `include` extended with `../shared-ui/src/primitives` so editor
    IntelliSense still works
- `app/web/vite.config.ts`: matching regex alias added before the
  catch-all `@`
- `app/web/package.json`: `"shared-ui": "workspace:*"` declared

## Source-level adjustments

5 files inside app/web/src/components/ used relative imports
(`./ui/tooltip`, `./ui/button`, etc.) to reach their old siblings.
After the move those siblings no longer live in web's component
tree. Switched all 7 import lines to the alias path
(`@/components/ui/...`):

- main.tsx (1 import)
- AppShell.tsx (1)
- CommandPalette.tsx (1)
- SidebarNav.tsx (1)
- Topbar.tsx (3)

No source-level edit was needed inside the moved primitives
themselves — they already used `@/lib/utils` and `@/stores/theme`,
which the existing alias chain (set up in A2) routes into shared/.

## Subtle gotcha worth noting in PR description

`shared-ui` initially didn't declare `@types/react` as a devDependency.
That looked harmless because `peerDependencies` covers `react` itself.
But cross-package TS type flow lost forwardRef's generic specialization,
which surfaced as 12 spurious `TS7006: Parameter 'e' implicitly has
an 'any' type` errors in web's pages (Login, Plugins). Adding
@types/react + @types/react-dom to shared-ui's devDependencies
restored the type flow and the 12 errors disappeared.

## Verification

- pnpm install --frozen-lockfile — 3 workspaces resolve clean
- pnpm --filter web build — 2728 modules transformed, chunks
  identical, internal/web/dist still populated
- go build ./cmd/opendray — go:embed resolves
- Existing main-on-web ESLint backlog unaffected (still 80
  errors + 8 warnings in web/components/, all pre-existing,
  not addressed here)
