+++ app/shared-ui/tsconfig.json
@@ -1,19 +1,17 @@
 {
   "compilerOptions": {
-    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "target": "es2023",
     "lib": ["ES2023", "DOM", "DOM.Iterable"],
     "module": "esnext",
-    "types": ["vite/client"],
-    "skipLibCheck": true,
-
     "moduleResolution": "bundler",
+    "skipLibCheck": true,
     "allowImportingTsExtensions": true,
     "verbatimModuleSyntax": true,
     "moduleDetection": "force",
     "noEmit": true,
     "jsx": "react-jsx",
 
+    "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "erasableSyntaxOnly": true,
@@ -21,10 +19,8 @@
 
     "paths": {
       "@/lib/*": ["../shared/src/lib/*"],
-      "@/stores/*": ["../shared/src/stores/*"],
-      "@/components/ui/*": ["../shared-ui/src/primitives/*"],
-      "@/*": ["./src/*"]
+      "@/stores/*": ["../shared/src/stores/*"]
     }
   },
-  "include": ["src", "../shared/src", "../shared-ui/src/primitives"]
+  "include": ["src"]
 }
