+++ app/web/package.json
@@ -38,6 +38,7 @@
     "react-markdown": "^10.1.0",
     "remark-gfm": "^4.0.1",
     "shared": "workspace:*",
+    "shared-ui": "workspace:*",
     "simple-icons": "^16.18.1",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.5.0",
