+++ app/web/src/main.tsx
@@ -5,7 +5,7 @@ import { RouterProvider } from '@tanstack/react-router'
 
 import './index.css'
 import { router } from './router'
-import { Toaster } from './components/ui/sonner'
+import { Toaster } from '@/components/ui/sonner'
 import '@/stores/theme' // side-effect: apply persisted theme
 
 const queryClient = new QueryClient({
